Filling for food and drink is now float not int.

+++ src/main/java/org/terasology/hunger/component/DrinkComponent.java
@@ -18,5 +18,5 @@ package org.terasology.hunger.component;
 import org.terasology.entitySystem.Component;
 
 public class DrinkComponent implements Component {
-    public int filling;
+    public float filling;
 }
